add class Heap and delete other non-used functional data structures
